+++ R/zzz.R
@@ -10,6 +10,7 @@
 
   op_bpr[[improved_annotation_option()]] <- FALSE
   op_bpr[[variable_uuid_option()]] <- FALSE
+  op_bpr[[empty_blueprints_dir_option()]] <- FALSE
 
   toset <- !(names(op_bpr) %in% names(op))
   if (any(toset)) options(op_bpr[toset])